EES-7002 - reorganising Screener-related services and records / classes.
