+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/Interfaces/Screener/IDataSetScreenerClient.cs
@@ -1,5 +1,7 @@
 using GovUk.Education.ExploreEducationStatistics.Admin.Requests;
+using GovUk.Education.ExploreEducationStatistics.Admin.Responses.Screener;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Screener;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces.Screener;
