+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Database/ContentDbContext.cs
@@ -6,6 +6,7 @@ using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Model.Chart;
 using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
 using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Screener;
 using GovUk.Education.ExploreEducationStatistics.Common.Utils;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
@@ -165,10 +166,10 @@ public class ContentDbContext : DbContext
 
         modelBuilder
             .Entity<DataSetUpload>()
-            .Property(upload => upload.ScreeningProgress)
+            .Property(upload => upload.ScreenerProgress)
             .HasConversion(
                 r => JsonSerializer.Serialize(r, (JsonSerializerOptions)null),
-                r => JsonSerializer.Deserialize<DataSetScreeningProgress>(r, (JsonSerializerOptions)null)
+                r => JsonSerializer.Deserialize<DataSetScreenerProgress>(r, (JsonSerializerOptions)null)
             );
 
         modelBuilder
